+++ test/Bond.t.sol
@@ -52,7 +52,7 @@ contract BondTest is Test {
         reg = new MandateRegistry();
         anchorLog = new AnchorLog(reg);
         mock = new MockFdc();
-        bond = new Bond(reg, anchorLog, IFdcVerification(address(mock)));
+        bond = new Bond(reg, anchorLog, IFdcVerification(address(mock)), 24 hours, 1 hours);
         reg.setBond(address(bond));
 
         vm.warp(1_800_000_000);